Crate elp

Structs

ELBRecord
The product of parsing a single AWS ELB log record.

ParsingErrors
The result of a failed attempt to parse an ELB record.

Enums

ELBRecordField
DON'T USE THIS IN YOUR CODE!!! This is really an implementation detail and shouldn't be exposed as part of the public API. Unfortunately it must be made public in order to implement the Index trait. I could use the newtype pattern but the newtype pattern forces another level of indirection with no gain besides reducing the exposure a little. I hope that in the future we'll be able to implement public methods without having to expose, what should be, private details. Update 02/14/2016: This behaviour has been changed in 1.7.0 nightly. This will be made private as soon as 1.7.0 is released.

ELBRecordParsingError
Specific parsing errors that are returned as part of the ParsingErrors::errors collection.

Functions

file_list
A utility method for retrieving all of the paths to ELB log files in a directory.

parse_record
Attempt to parse a single string into an ELB record.

process_files
Attempt to parse every ELB record in every file in filenames and pass the results to the record_handler.

Type Definitions

ParsingResult
The result of an attempt to parse an ELB record.