Enum elp::ELBRecordField

pub enum ELBRecordField {
    Timestamp,
    ELBName,
    ClientAddress,
    BackendAddress,
    RequestProcessingTime,
    BackendProcessingTime,
    ResponseProcessingTime,
    ELBStatusCode,
    BackendStatusCode,
    ReceivedBytes,
    SentBytes,
    RequestMethod,
    RequestURL,
    RequestHTTPVersion,
    UserAgent,
    SSLCipher,
    SSLProtocol,
}

DON'T USE THIS IN YOUR CODE!!! This is really an implementation detail and shouldn't be exposed as part of the public API. Unfortunately it must be made public in order to implement the Index trait. I could use the newtype pattern but the newtype pattern forces another level of indirection with no gain besides reducing the exposure a little. I hope that in the future we'll be able to implement public methods without having to expose, what should be, private details. Update 02/14/2016: This behaviour has been changed in 1.7.0 nightly. This will be made private as soon as 1.7.0 is released.

Variants

Timestamp

ELBName

ClientAddress

BackendAddress

RequestProcessingTime

BackendProcessingTime

ResponseProcessingTime

ELBStatusCode

BackendStatusCode

ReceivedBytes

SentBytes

RequestMethod

RequestURL

RequestHTTPVersion

UserAgent

SSLCipher

SSLProtocol