Function elp::file_list

pub fn file_list(dir: &Path, filenames: &mut Vec<DirEntry>) -> Result<usize, Error>

A utility method for retrieving all of the paths to ELB log files in a directory.

If the user uses the AWS S3 sync tool to download their AWS ELB logs to a local disk the files will be in a very specific directory hierarchy. This utility will read the paths of the files, recursively searching a root specified by the user, and append the paths to the Vec<DirEntry>, also provided by the user.

dir: The directory from which the paths of the ELB log files will be procured.

filenames: A Vec to which the paths of the ELB log files will be written.