Struct elp::ParsingErrors

pub struct ParsingErrors<'a> {
    pub record: &'a str,
    pub errors: Vec<ELBRecordParsingError>,
}

The result of a failed attempt to parse an ELB record.

It is very possible that multiple fields of a record are not parsable. An attempt is made to parse all of the fields of an ELB record. An error is returned for each field that was not parsable to make it clear what fields of the record were faulty and allow the user to decide how to handle the failure.

Fields
record: &'a str

The raw record.

errors: Vec<ELBRecordParsingError>

A collection of parsing errors such as fields that could not be parsed or a failure to open an ELB log file.